requirements.txt
torch
torchvision
pillow
pdf2image
beautifulsoup4
requests
streamlit
pandas
scikit-learn
matplotlib
tqdm
streamlit
pymupdf
Pillow
requests